feat: implement core service management logic and repair workflow with asynchronous repository operations

+++ src/main/java/tr/cabro/servicio/application/listeners/ServiceEditListener.java
@@ -3,10 +3,12 @@ package tr.cabro.servicio.application.listeners;
 import tr.cabro.servicio.model.AddedPart;
 import tr.cabro.servicio.model.Process;
 
+import java.math.BigDecimal;
+
 public interface ServiceEditListener {
     // Mevcut metotlar
-    void onPartChange(double totalPartPrice);
-    void onProcessAdded(String name, double price);
+    void onPartChange(BigDecimal totalPartPrice);
+    void onProcessAdded(String name, BigDecimal price);
     void onProcessAdded(Process process);
     void onStatusChanged(String status);
